% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_counties}
\alias{tn_counties}
\title{Dataframe of counties and their health regions}
\format{
An object of class \code{data.frame} with 96 rows and 3 columns.
}
\usage{
tn_counties
}
\description{
Dataframe of counties and their health regions
}
\keyword{datasets}
