% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_county_group}
\alias{tn_county_group}
\title{Merge county shapefiles by a group definition}
\usage{
tn_county_group(group_df, crs = NA, buffer = 1e-06)
}
\arguments{
\item{group_df}{Dataframe with exactly 2 columns. One must be the county name or FIPS code (column name must have 'county' or 'fips' in it). The other must be the grouping variable.}

\item{crs}{EPSG code for the desired output CRS. NA defaults to the CRS for tn_county_shapefile (6576)}

\item{buffer}{Amount to buffer each county before merging.}
}
\value{
Shapefile for the merged counties
}
\description{
When you provide a dataframe of counties and a grouping variable, this function will create a merged shapefile.
You can specify a CRS to use, but the default is the CRS for tn_county_shapefiles.
}
