% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_county_shapefiles}
\alias{tn_county_shapefiles}
\title{Shapefiles for each county in TN}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 95 rows and 60 columns.
}
\usage{
tn_county_shapefiles
}
\description{
Shapefiles and associated data for counties in TN
}
\keyword{datasets}
