% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_geocode_addresses}
\alias{tn_geocode_addresses}
\title{Function to interface with the TN geocoder.}
\usage{
tn_geocode_addresses(
  df,
  match_on = c(Address = "Address", City = "City", State = "State", Postal = "Postal"),
  return_fields = c("Score", "Match_addr", "County", "X", "Y"),
  exclude_default_fields = T,
  locator = "TN_COMPOSITE"
)
}
\arguments{
\item{df}{dataframe to use}

\item{match_on}{Named character vector of columns in the dataframe that map to geocoded elements (geocoder_name=df_name).
Or an unnamed vector of columns that are valid geocoder inputs.
Or a partially named vector where columns that are not valid geocoder inputs are named as valid inputs.
Valid geocoder inputs can be found using tn_api_inputs(). You may also specify a single column with a combined address using 'SingleLine'='column_name'.
This will cause any other input fields to be ignored.}

\item{return_fields}{Character vector of valid fields to return. Valid geocoder inputs can be found using tn_api_outputs().
To return all possible fields, use '*','','All', or NA. To return the default minimum fields, use 'None'.
Invalid fields generate a warning and will be ignored.}

\item{exclude_default_fields}{T/F Should default geocoding fields not listed in return_fields be excluded from the output?
Default fields include: Address, X, Y, and Score. This parameter is ignored if return_fields is '*','','All','None', or NA.}
}
\value{
Dataframe with added columns from geocoder
}
\description{
Interface to https://tnmap.tn.gov/arcgis/rest/services/LOCATORS/TN_COMPOSITE/GeocodeServer/geocodeAddresses
}
