% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_geocoder_specialcases}
\alias{tn_geocoder_specialcases}
\title{Special input/outputs handled by tn_geocode_addresses()}
\usage{
tn_geocoder_specialcases()
}
\value{
dataframe of aliases and the official names they correspond to.
}
\description{
These names are not recognized by the geocoding api, but are handled by tn_geocode_addresses().
}
