% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_region_shapefiles}
\alias{tn_region_shapefiles}
\title{Shapefiles for each region in TN}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 51 columns.
}
\usage{
tn_region_shapefiles
}
\description{
Shapefiles and associated data for regions in TN
}
\keyword{datasets}
