% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_validate_county}
\alias{tn_validate_county}
\title{Standardize TN county names}
\usage{
tn_validate_county(county)
}
\arguments{
\item{county}{vector of counties in TN}
}
\value{
vector of standardized county names
}
\description{
Removes " county" and standardizes capitalization and spacing for TN counties.
}
