% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_zipcode_shapefiles}
\alias{tn_zipcode_shapefiles}
\title{Shapefiles for each zipcode in TN}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 609 rows and 53 columns.
}
\usage{
tn_zipcode_shapefiles
}
\description{
Shapefiles and associated data for zipcodes in TN
}
\keyword{datasets}
