% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_county_to_region}
\alias{tn_county_to_region}
\title{Convert TN counties to health regions}
\usage{
tn_county_to_region(county)
}
\arguments{
\item{county}{vector of counties in TN}
}
\value{
vector of health regions
}
\description{
Convert TN counties to health regions
}
