% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{tn_geocode_vector}
\alias{tn_geocode_vector}
\title{Function to interface with the TN geocoder.}
\usage{
tn_geocode_vector(
  x,
  return_fields = c("Score", "Match_addr", "County", "X", "Y"),
  exclude_default_fields = T
)
}
\arguments{
\item{x}{A vector of addresses}

\item{return_fields}{Character vector of valid fields to return. Valid geocoder inputs can be found using tn_api_outputs().
To return all possible fields, use '*','','All', or NA. To return the default minimum fields, use 'None'.
Invalid fields generate a warning and will be ignored.}

\item{exclude_default_fields}{T/F Should default geocoding fields not listed in return_fields be excluded from the output?
Default fields include: Address, X, Y, and Score. This parameter is ignored if return_fields is '*','','All','None', or NA.}
}
\value{
Dataframe with added columns from geocoder
}
\description{
A wrapper around tn_geocode_addresses() that takes a vector of addresses as input
}
