% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{validate_county_group}
\alias{validate_county_group}
\title{Validate a county group dataframe}
\usage{
validate_county_group(group_df)
}
\arguments{
\item{county}{vector of counties in TN}
}
\value{
vector of standardized county names
}
\description{
Check that a data frame meets the requirements to function in tn_county_group.
Dataframe with exactly 2 columns. One must be the county name or FIPS code (column name must have 'county' or 'fips' in it). The other must be the grouping variable.
}
